Support tracking by using WebSockets
This adds support to use session tracking by using a WebSocket connection, if
the TV device supports it. Otherwise use the previously used tracking pixel
method as a fallback.

The WebSocket method relies on the protocol level heartbeat of the WebSocket
protocol, that means it relies on the server to periodically send ping packets
in order to keep the conneciton open. Those ping packets are transparent to
the client side, i.e. the HbbTV app is not aware of those pings.

Should the WebSocket connection be closed by the server or by some error, the
tracking script will continuously try and create a new WebSocket connection
until successful. Here the same error back-off algorithm is uses as previously
with the tracking pixel implementation so that the backend will not get
swamped in requests.

The template for the tracking script introduces a new variable for the
WebSocket connection URL, i.e. {{WEBSOCKET_URL}} that should define
protocol scheme, hostname and port (port optional), without slash in the end,
example:

	ws://session-cl.tvping.com

The full URL for the WebSocket connection as opened by the tracking script
would then e.g. be:

	ws://session-cl.tvping.com/$CHANNEL_ID?$TIMESTAMP

Where $CHANNEL_ID is the current numeric channel id and $TIMESTAMP is the same
client side UNIX timestamp in milliseconds as previously used by the tracking
pixel URL.